requirements.txt
python-telegram-bot==22.6
python-telegram-bot
google-generativeai